pdk: ti: drv: sciclient: src: sciserver_makefile: Set the versions fixed to prevent auto-generation with empty string
The scm version (of the PDK) cannot be auto-detected from git tags in
the Phytec fork of the SDK, so we are setting the version fixed here and
add a '-phy' to indicate the Phytec fork. Appending a git hash/branch
name/Phytec tag may be done in the generation script in a separate commit.

Signed-off-by: Steffen Hemer <s.hemer@phytec.de>