debug probe: add openocd script for XDS110
The TI XDS110 debug probe is supported by Openocd. Add an Openocd script
that allows connecting with the target

Connect with:
  openocd -f j721s2_openocd.cfg

And attach with:
  gdb-multiarch
  target extended-remote localhost:3336

To one of the following Targets:
  j721s2.cpu.sysctrl on 3333
  j721s2.cpu.a72.0 on 3334
  j721s2.cpu.a72.1 on 3335
  j721s2.cpu.mcu_r5.0 on 3336
  j721s2.cpu.mcu_r5.1 on 3337
  j721s2.cpu.main0_r5.0 on 3338
  j721s2.cpu.main0_r5.1 on 3339
  j721s2.cpu.main1_r5.0 on 3340
  j721s2.cpu.main1_r5.1 on 3341
  j721s2.cpu.gp_mcu on 3342

TODO:
Currently a monitor reset halt command always resets the whole SoC
(pulling reset line). There should be an option to make a soft reset for
individual cores, that e.g. just set the program counter. Research such
an option and test.

Signed-off-by: Jonas Remmert <j.remmert@phytec.de>