ipc_test sample: exit firmware if shutdown was requested
When stopping a remote core, the Linux driver waits until the core
confirms a shutdown of the firmware. This shutdown is requested by e.g.
Linxu via IPC. This commit checks for the shutdown command, sends a
shutdown confirm and exits the firmware.

The following error was triggered when a stop of a remote core has been
attempted:

root@phyboard-izar-am68x-2:~# echo stop > /sys/class/remoteproc/remoteproc0/state
[   73.799589] k3-dsp-rproc 64800000.dsp: k3_dsp_rproc_stop: timedout waiting for rproc completion event
[   73.808812] remoteproc remoteproc0: can't stop rproc: -16
-sh: echo: write error: Device or resource busy

Signed-off-by: Jonas Remmert <j.remmert@phytec.de>