ipc_test: Enable debug prints
The ipc_test example serves only as an example to demonstrate multicore
communication therefore it makes sense to enable debug per default.

Per default logging to the tracebuffer in memory is enabled. The logging
outputs can be viewed with:

cat /sys/kernel/debug/remoteproc/remoteproc<n>/trace0 where 'n' is the
number of the remoteproc interface for the related core.

Signed-off-by: Jonas Remmert <j.remmert@phytec.de>